pyproject.toml
[tool.poetry]
name = "pycord-tf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"
asyncio = "^3.4.3"
pysqlite3 = "^0.5.3"
websockets = "^12.0"
mariadb = "^1.1.10"
discord = "^2.3.2"
schedule = "^1.2.2"
python-levenshtein = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
